+++ UnitTest/NCLBoardUnitTest.cpp
@@ -392,7 +392,7 @@ namespace UnitTest
 				NCLBoard board1(3, 1, 8);
 				NCLBoardTraverser t(&board1, true);
 				ContinuousNumber c = t.GetTotalContinuousNumber();
-				Assert::AreEqual(board1.GetTotalContinuousNumber(true), (unsigned long)c.row);
+				Assert::AreEqual(board1.GetTotalContinuousNumber(true), (unsigned long long)c.row);
 			}
 		}
 
@@ -404,7 +404,7 @@ namespace UnitTest
 				NCLBoard board1(3, 1, 12);
 				NCLBoardTraverser t(&board1, true);
 				ContinuousNumber c = t.GetTotalContinuousNumber();
-				Assert::AreEqual(board1.GetTotalContinuousNumber(true), (unsigned long)c.row);
+				Assert::AreEqual(board1.GetTotalContinuousNumber(true), (unsigned long long)c.row);
 			}
 		}
 
@@ -416,7 +416,7 @@ namespace UnitTest
 				NCLBoard board1(3, 10, 30);
 				NCLBoardTraverser t(&board1, true);
 				ContinuousNumber c = t.GetTotalContinuousNumber();
-				Assert::AreEqual(board1.GetTotalContinuousNumber(true), (unsigned long)c.row);
+				Assert::AreEqual(board1.GetTotalContinuousNumber(true), (unsigned long long)c.row);
 			}
 		}
 	};
